Create 06 uniqueRows.cpp
Since we cannot use unordered_map<vector<int>,int>, we have to represent our row as a String. Then in unordered map we can check if we see some row for the first time and if yes we push it to output. Time complexity : O(Row * Col) because we have to go through every element in matrix.