Create 11 relativeSort.cpp
First we fill unordered map with elements from arr1 and we count their frequency. That is in time O(n).
Then we iterate through arr2 and we are looking them up in our unordered map. If we find it there we push_back the numbers to output. That is in time max O(n). Then we fill up our array notInArr2. That is in time O(n). Then we sort that array. It is in time O(n log n). And then we push back numbers from notInArr2 to output.
Time complexity of this all is O(n log n)