Create 03 convertToRomanNumber.cpp
We create 2 helping arrays of size 13. One is filled with roman numbers and the second one is filled with arabic numbers representing those roman numbers in the first array. Then in a for and while cycle we are constructing our roman number and we are decreasing N. If N is zero we break out of cycle.
Time complexity is close to constant but it is not so time complexity is : O(log10 N)